feature: over/under clocking
- You can now select settings before starting the game
- FC and MD over/underclocking
- All over/underclocking can take a number as a param (select box)
- When adding a ROM, auto-select it

+++ higan/emulator/interface.hpp
@@ -6,7 +6,7 @@ struct Interface {
   virtual ~Interface() = default;
 
   //configuration
-  virtual auto configure(string name, uint value) -> void {}
+  virtual auto configure(string name, double value) -> void {}
   
   //information
   virtual auto name() -> string { return {}; }
