+++ higan/md/interface/interface.hpp
@@ -5,7 +5,7 @@ namespace higan::MegaDrive {
 extern Interface* interface;
 
 struct MegaDriveInterface : Interface {
-  auto configure(string, uint) -> void override;
+  auto configure(string, double) -> void override;
   auto name() -> string override { return "Mega Drive"; }
   auto game() -> string override;
