+++ higan/target-web/platform/platform.hpp
@@ -63,7 +63,7 @@ struct WebPlatform : higan::Platform {
         auto unload() -> void;
 
         auto run() -> void;
-        auto configure(string name, uint value) -> void;
+        auto configure(string name, double value) -> void;
         auto setVolume(uint volume) -> void;
         auto setMute(bool mute) -> void;
